[WIP] Support tensors as Dict keys
This prepares the PR where we implement sets in terms of dicts.

The Hashable variable is opaque on purpose so that it fails hard if
if it unadvertedly leaks back into user code.

We also found and fixed a number of latent bugs and inconsistencies
in the way dynamo checked what can be a dict key.

Fixes https://github.com/pytorch/pytorch/issues/107595
Fixes https://github.com/pytorch/pytorch/issues/111603

ghstack-source-id: a8d00159facc0e8b37f01a422a89b9fc08b76336
Pull Request resolved: https://github.com/pytorch/pytorch/pull/111196

Fix import