[C10D] FlightRecorder don't compute duration inside dump()
Computing duration can lead to a hang due to calling cudaEventQuery when
the cuda driver queue is full.

We don't ever want dump() api to hang, since we might want dump to help
debug a hang.  Watchdog may already hang due to cuda event queries, and
we have the monitor to kill and dump if watchdog hangs.

Perhaps we should use a separate thread for computing durations..

And this still doesn't fix test_trace_while_stuck?

ghstack-source-id: 2f78fa23140217a13e5b29da20aca7ada899a121
Pull Request resolved: https://github.com/pytorch/pytorch/pull/116909